moved constants to utils
